docs: provide descriptions for commands exposed via Composer
Composer supports a "scripts-descriptions" setting, which is an object
where keys are the script names, and their values the descriptions.
These are then used during a `list` operation (or running `composer`
without arguments) to detail the purpose of the script.

This patch provides descriptions for all commands except
`post-create-project-command`, which is removed following project
creation.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>